Moved WordGroupLike trait to the separate file
